encapsulate recorder
